Code Revision This is a revision. Due to challenges with the way the client, scheduler and local worker where initial implemented, I had to do a code revision. The reasons why will be explained as I type.
I will start with each java class and what they do, in it you will find
the the reason for each change.

The Client Class
There was an issue with this class, one major issue was that clients
where not getting results of task they hard submited. Secondly, We were
required to add some form of unquie id to the tasks. The revisied class
allows clients to get the results from the scheduler as soon as workers
are done processing their jobs. This works for both local jobs and
remote jobs. Also some form of identification was added to each task by
concatenating the client hose address with the index of the task in the
task list plus the sleep task value. Client send and receive tasks. 

The Scheduler Class
The class was designed without multhreading. There was no way multiple
clients would connect to the scheduler at the same time. So we had to
split the scheduler into two various classes to seperate concerns. Both
of them answer the extra credit part where we are to send them back to
the client in a batch.

The LocalScheduler and LocalWorker Class
It hands the relationship with sufi's local worker. The local worker
sends back result to a queue as directed in the assignment. The local
worker was modified to provide this feature.

The RemoteScheduler Class
Handles relationship with the remote worker. When the client sends the
batch jobs to the scheduler, the scheduler depending on if it was remote
worker that was selected when it was ran spawns a remote scheduler
thread which handles communciation with remote client. When the remote
scheduler sends jobs to the queue, it then creates a thread that handles
getting messages for the client, a sqs queue is created for each client
so that workers can send result to their queue when done and the
scheduler will then peak it, check if it is for the client then checks
if it has not already added it to the list before it adds it to the
list. When the number of received task is equal to the number of
submitted task by the client it sends the task to the client, this also
answers the extra credit part of sending batch jobs back to the client.
Also the remote worker sends task to queue in batchs of 10 which is the
maximum sqs allows. How this works is that when we receive the number of
task from client, we check the size, if its less than or equal to 10 we
send it as a batch job using sqs api else, we loop through the task,
when find the modulo of the iteration counter by 10, it its zero then we
send the tasks saved so far as a batch and reinstiate the batch object.
If it happens that after server modulos, we are left with numbers that
are less than 10 and the loop has finished iterating, it is assumed that
the content is less than 10 and we send them as a batch. This whole
process answeres another extra credit about sending batch jobs to the
queue for workers to process

Remote Worker and RemoteSleepTask Class
Is the remote version of the local worker, it spawns a new that pools
every 20 seconds if the is a message in the queue. If after 20 seconds
it polls and no messages are present it goes to sleep. some times the
sqs delays in make a message available that is why there is a sleep and
20 seconds polling. if messages are found in the queue, client gets the
messages, again this is done in batch, which answers another extra
credit question. A pool of thread is created based on the size of the
messages. This answers another extra credit problem. Each tread runs a
RemoteSleepTask. This class checks if the dynamodb has been created, if
not it creates its. It checks the database to see if the task given by
the remote worker to handle is not in the database so it can work on it.
The dynamodb is represented as numbers which consist of the ipaddress of
the client concatenated with the index of the task in the task list.
When the task is not in the database the class inserts it into the
database and then executes the sleep task function. After it is done, it
sends a message to the queue of the client that owns the task. 

DdbRemoteService, SleepTaskItem and SqsRemoteService are wrappers around
Dynamodb and SQS respectively. The DdbRemoteService allows us perform
dynamodb functions easily, sleepTaskItem is a Dynamodb object that is
used to represent our entity. The SqsRemoteService makes it easy to run
SQS apis.

We have removed every form of handcoded values. Every value is now
gotten from the argument passed to the applications. Since we are to run
it on amazon a jar file has been created for each section: the client,
the scheduler and the worker(remote). To run, run in the following order
1. java -jar Scheduler.jar -s <PORT> [if you want local add -lw <NUM of
threads to run> else if you want remote worker -rw]
2. you can run the client or worker in any other, for client
java -jar Client.jar -s <IP_ADDRESS:PORT> -w <PATH_TO_WORKLOAD_FILE>
for worker : java -jar Worker.jar -i <TIME_SEC>
The arguments parsed and their meaning is as it is in the assignment.